controller

+++ src/main/java/com/burakejder/entities/OrderItem.java
@@ -9,7 +9,7 @@ public class OrderItem {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer orderItemId;
+    private Long orderItemId;
 
 
     @ManyToOne
